Сделал заготовку для лоадинг скрина. Пока лодырь работает в нашем же потоке, она бесполезна, но потом будет норм.

+++ src/com/freedom/gameObjects/Robot.java
@@ -94,8 +94,8 @@ public class Robot extends Stuff implements Moveable {
 
 
 	public void recalibrate() {
-		x = (int) Math.round(x);
-		y = (int) Math.round(y);
+		x = Math.round(x);
+		y = Math.round(y);
 
 	}
